test: MET-2190 refactor folder
